+++ src/pageEditor/extensionPoints/panel.ts
@@ -30,7 +30,7 @@ import {
   removeEmptyValues,
   selectIsAvailable,
 } from "@/pageEditor/extensionPoints/base";
-import { omitEditorMetadata } from "./normalizePipeline";
+import { omitEditorMetadata } from "./pipelineMapping";
 import { ExtensionPointConfig } from "@/extensionPoints/types";
 import {
   PanelConfig,
