+++ src/pageEditor/extensionPoints/sidebar.tsx
@@ -31,7 +31,7 @@ import {
   removeEmptyValues,
   selectIsAvailable,
 } from "@/pageEditor/extensionPoints/base";
-import { omitEditorMetadata } from "./normalizePipeline";
+import { omitEditorMetadata } from "./pipelineMapping";
 import { ExtensionPointConfig } from "@/extensionPoints/types";
 import {
   SidebarConfig,
